requirements.txt
requests == 2.32.3
fastapi == 0.115.12
uvicorn == 0.34.2
python-dotenv == 1.1.0
sqlalchemy == 2.0.39
psycopg2-binary == 2.9.10
line-bot-sdk == 3.17.1
